Extract ComparePopolo class to separate file
